move col implementations to own module
