+++ crates/argmin-math/src/faer_m_0_21/col/dot.rs
@@ -1,0 +1,52 @@
+use crate::ArgminDot;
+use faer::prelude::*;
+use faer_traits::ComplexField;
+
+/// contains implementations for applying matrices to column vectors.
+mod matrix_column_multiplication {
+    use super::*;
+
+    /// MatRef . ColRef -> Col
+    impl<E: ComplexField> ArgminDot<ColRef<'_, E>, Col<E>> for MatRef<'_, E> {
+        #[inline]
+        fn dot(&self, other: &ColRef<'_, E>) -> Col<E> {
+            self * other
+        }
+    }
+
+    /// Mat . Col -> Col
+    impl<E: ComplexField> ArgminDot<Col<E>, Col<E>> for Mat<E> {
+        #[inline]
+        fn dot(&self, other: &Col<E>) -> Col<E> {
+            <_ as ArgminDot<_, _>>::dot(&self.as_ref(), &other.as_ref())
+        }
+    }
+}
+
+/// contains implementations for the scalar product of two column vectors of
+/// the same length. This is v^H . u for two column vectors v,u.
+mod scalar_product {
+    use super::*;
+    use faer_traits::Conjugate;
+
+    /// ColRef . ColRef -> Col
+    impl<E: ComplexField + Conjugate<Conj = E>> ArgminDot<ColRef<'_, E>, E> for ColRef<'_, E> {
+        #[inline]
+        fn dot(&self, other: &ColRef<'_, E>) -> E {
+            assert_eq!(
+                self.nrows(),
+                other.nrows(),
+                "vectors for dot product must have same number of elements"
+            );
+            self.conjugate().transpose() * other
+        }
+    }
+
+    /// Col . Col -> Col
+    impl<E: ComplexField + Conjugate<Conj = E>> ArgminDot<Col<E>, E> for Col<E> {
+        #[inline]
+        fn dot(&self, other: &Col<E>) -> E {
+            <_ as ArgminDot<_, _>>::dot(&self.as_ref(), &other.as_ref())
+        }
+    }
+}
