+++ crates/argmin-math/src/faer_m_0_21/col/l2_norm.rs
@@ -1,14 +1,14 @@
 use crate::ArgminL2Norm;
-use faer::{Mat, MatRef};
+use faer::{Col, ColRef};
 use faer_traits::ComplexField;
 
-impl<E: ComplexField> ArgminL2Norm<E::Real> for MatRef<'_, E> {
+impl<E: ComplexField> ArgminL2Norm<E::Real> for ColRef<'_, E> {
     fn l2_norm(&self) -> E::Real {
         self.norm_l2()
     }
 }
 
-impl<E: ComplexField> ArgminL2Norm<E::Real> for Mat<E> {
+impl<E: ComplexField> ArgminL2Norm<E::Real> for Col<E> {
     fn l2_norm(&self) -> E::Real {
         self.norm_l2()
     }
