+++ crates/argmin-math/src/faer_m_0_21/col/mod.rs
@@ -1,0 +1,4 @@
+mod dot;
+mod l2_norm;
+mod mul;
+mod sub;
